Added file to generate TF dataset
I haven't fully tested this, but the TF dataset generator should accept
the directory structure of the dataset we currently have and feed it
straight into the model.fit() function.
It should be called separately on the train and test directories.

Also added Settings.py to consolidate declarations and variables that
will likely be used in more than one place.
This will help keep the code DRY and also reduce errors if values must
change.